requirements.txt
openai>=1.0.0
jinja2>=3.1.0
python-dotenv>=1.0.0
markdown>=3.4.0
weasyprint>=60.0
pandas>=2.0.0
pyyaml>=6.0
click>=8.1.0
rich>=13.0.0
python-dateutil>=2.8.0
flask>=2.3.0
